Add fail2ban and more options

+++ oci-setup.sh
@@ -6,15 +6,19 @@ if ((${EUID:-0} || "$(id -u)")); then
 fi
 
 # Set default values for options and variables
-do_update=false
+force_confirm_needed=false
 welcome_choice=""
 confirm_needed=true
+install_fail2ban=""
+add_missing_shh_entry=""
+quiet=false
+ssh_port=""
 
 # Parse command-line options
-while getopts "yw:" opt; do
+while getopts "yw:fs:mq" opt; do
   case "${opt}" in
     y)
-      do_update=true
+      force_confirm_needed=true
       ;;
     w)
       if [[ "${OPTARG}" == "p" || "${OPTARG}" == "c" || "${OPTARG}" == "t" ]]; then
@@ -24,11 +28,28 @@ while getopts "yw:" opt; do
         exit 1
       fi
       ;;
+    f)
+      install_fail2ban="y"
+      ;;
+    s)
+     if [[ "${OPTARG}"  =~ ^[0-9]+$ ]]; then
+        ssh_port="${OPTARG}"
+      else
+        echo "Invalid option argument for -s: ${OPTARG}" >&2
+        exit 1
+      fi
+      ;;
+    m)
+      add_missing_shh_entry="y"
+      ;;
+    q)
+      quiet=true
+      ;;
   esac
 done
 
 # Check if confirmation is needed
-if $do_update; then
+if $force_confirm_needed; then
   confirm_needed=false
 fi
 
@@ -85,8 +106,13 @@ sudo dpkg-reconfigure -f noninteractive unattended-upgrades
 sudo sed -i 's/^.*printf/# &/' /etc/update-motd.d/10-help-text
 
 ## Add custom message
-if [ -z "$welcome_choice" ]; then
-  read -p "Do you want to use the production or test custom welcome message? (p/c/t) " welcome_choice
+if ! $quiet; then
+    while [[ ! "$welcome_choice" =~ ^(p|c|d|)$ ]]; do
+        read -p "Do you want to use the production or test custom welcome message? (p/c/t) " welcome_choice
+        if [[ ! "$welcome_choice" =~ ^(p|c|d|)$ ]]; then
+            echo "Invalid input. Please enter 'p', 'c', 't', or leave it blank."
+        fi
+    done
 fi
 
 if [ "$welcome_choice" == "p" ]; then
@@ -130,6 +156,15 @@ EOF
     sudo chmod +x /etc/update-motd.d/99-custom-welcome
 fi
 
+if ! $quiet; then
+    while [[ ! "$add_missing_shh_entry" =~ ^(y|n)$ ]]; do
+        read -p "Do you want to add missing ssh folders/files? (y/n): " add_missing_shh_entry
+        if [[ ! "$add_missing_shh_entry" =~ ^(y|n)$ ]]; then
+            echo "Invalid input. Please enter 'y' or 'n'."
+        fi
+    done
+fi
+
 # Get all system usernames (getting uid between 1000 and 1100, excluding nobody)
 usernames=$(getent passwd | awk -F: '{if ($3>=1000 && $3<=1100 && $1!="nobody") print $1}')
 
@@ -155,22 +190,52 @@ do
     fi
 done
 
-## Increase ssh timeout
-sudo sed -i "s/#ClientAliveInterval 0/ClientAliveInterval 600/g" /etc/ssh/sshd_config
-sudo sed -i "s/#ClientAliveCountMax 3/ClientAliveCountMax 6/g" /etc/ssh/sshd_config
-sudo sed -i "s/#Port 22/Port 2233/g" /etc/ssh/sshd_config
-
-sudo systemctl reload sshd
+if ! $quiet; then
+    while [[ ! "$ssh_port" =~ ^[0-9]+$ || -z "$ssh_port" ]]; do
+        read -p "Enter the SSH port number: " ssh_port
+        if [[ ! "$ssh_port" =~ ^[0-9]+$ ]]; then
+            echo "Invalid input. Please enter a valid SSH port number."
+        elif [[ -z "$ssh_port" ]]; then
+            break
+        fi
+    done
+fi
+if [[ -z "$ssh_port" ]]; then
+    echo "No port specified. Defaulting to port 2233."
+    ssh_port="2233"
+fi
 
 ## Install ufw
 sudo apt install ufw -y
 sudo ufw default deny incoming
 sudo ufw default allow outgoing
-sudo ufw allow 2233
+sudo ufw allow $ssh_port
 sudo ufw allow http
 sudo ufw allow https
 
 sudo ufw --force enable
+
+# Install and configure Fail2Ban
+if ! $quiet; then
+    while [[ ! "$install_fail2ban" =~ ^(y|n)$ ]]; do
+        read -p "Do you want to install fail2ban? (y/n): " install_fail2ban
+        if [[ ! "$install_fail2ban" =~ ^(y|n)$ ]]; then
+            echo "Invalid input. Please enter 'y' or 'n'."
+        fi
+    done
+fi
+
+if $install_fail2ban; then
+    sudo apt-get install fail2ban -y
+    sudo systemctl enable fail2ban --now
+fi
+
+## Increase ssh timeout
+sudo sed -i "s/#ClientAliveInterval 0/ClientAliveInterval 600/g" /etc/ssh/sshd_config
+sudo sed -i "s/#ClientAliveCountMax 3/ClientAliveCountMax 6/g" /etc/ssh/sshd_config
+sudo sed -i "s/#Port 22/Port $ssh_port/g" /etc/ssh/sshd_config
+
+sudo systemctl reload sshd
 
 ## Run tests
 echo -e "\n==== Run tests ===="
@@ -192,7 +257,7 @@ if ! systemctl status nginx | grep "active (running)"; then
     error=true
 fi
 
-if [ ! grep -q "^TLSv1" /etc/nginx/nginx.conf ]; then
+if ! grep -q "^TLSv1" /etc/nginx/nginx.conf; then
     echo "TLSv1 is not removed."
     error=true
 fi
@@ -220,8 +285,8 @@ if [ "$timeout" != "6" ]; then
 fi
 
 timeout=$(sudo grep "^Port" /etc/ssh/sshd_config | awk '{print $2}')
-if [ "$timeout" != "2233" ]; then
-    echo "SSH port is not set to 2233."
+if [ "$timeout" != $ssh_port ]; then
+    echo "SSH port is not set to $ssh_port."
     error=true
 fi
 
@@ -230,7 +295,7 @@ if ! grep -q "APT::Periodic::Unattended-Upgrade" /etc/apt/apt.conf.d/20auto-upgr
     error=true
 fi
 
-if [ ! -f /etc/update-motd.d/99-custom-welcome ]; then
+if [ ! -f /etc/update-motd.d/99-custom-welcome && "$welcome_choice" =~ ^(p|c|d|)$ ]; then
     echo "Custom welcome message is not added."
     error=true
 fi
